requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
openai==1.1.1
pydantic==1.10.8
python-dotenv==1.0.0
requests==2.31.0 